refactor: moved the functions to utils
