+++ src/utils.ts
@@ -1,5 +1,5 @@
-import { Server, TLSServeOptions } from "bun";
-import { Config, Route } from ".";
+import { Errorlike, Server, TLSServeOptions } from "bun";
+import { Config, Context, Controller, Group, Route } from ".";
 import fs from "fs";
 
 export function createPath(str: string) {
@@ -28,37 +28,37 @@ export enum HttpMethod {
 }
 
 export enum StatusCode {
-    OK = 200,
-    CREATED = 201,
-    ACCEPTED = 202,
-    NO_CONTENT = 204,
-    MOVED_PERMANENTLY = 301,
-    FOUND = 302,
-    SEE_OTHER = 303,
-    NOT_MODIFIED = 304,
-    TEMPORARY_REDIRECT = 307,
-    PERMANENT_REDIRECT = 308,
-    BAD_REQUEST = 400,
-    UNAUTHORIZED = 401,
-    PAYMENT_REQUIRED = 402,
-    FORBIDDEN = 403,
-    NOT_FOUND = 404,
-    METHOD_NOT_ALLOWED = 405,
-    NOT_ACCEPTABLE = 406,
-    REQUEST_TIMEOUT = 408,
-    CONFLICT = 409,
-    GONE = 410,
-    LENGTH_REQUIRED = 411,
-    PAYLOAD_TOO_LARGE = 413,
-    URI_TOO_LONG = 414,
-    UNSUPPORTED_MEDIA_TYPE = 415,
-    EXPECTATION_FAILED = 417,
-    IM_A_TEAPOT = 418,
-    TOO_MANY_REQUESTS = 429,
-    INTERNAL_SERVER_ERROR = 500,
-    NOT_IMPLEMENTED = 501,
-    BAD_GATEWAY = 502,
-    SERVICE_UNAVAILABLE = 503,
+  OK = 200,
+  CREATED = 201,
+  ACCEPTED = 202,
+  NO_CONTENT = 204,
+  MOVED_PERMANENTLY = 301,
+  FOUND = 302,
+  SEE_OTHER = 303,
+  NOT_MODIFIED = 304,
+  TEMPORARY_REDIRECT = 307,
+  PERMANENT_REDIRECT = 308,
+  BAD_REQUEST = 400,
+  UNAUTHORIZED = 401,
+  PAYMENT_REQUIRED = 402,
+  FORBIDDEN = 403,
+  NOT_FOUND = 404,
+  METHOD_NOT_ALLOWED = 405,
+  NOT_ACCEPTABLE = 406,
+  REQUEST_TIMEOUT = 408,
+  CONFLICT = 409,
+  GONE = 410,
+  LENGTH_REQUIRED = 411,
+  PAYLOAD_TOO_LARGE = 413,
+  URI_TOO_LONG = 414,
+  UNSUPPORTED_MEDIA_TYPE = 415,
+  EXPECTATION_FAILED = 417,
+  IM_A_TEAPOT = 418,
+  TOO_MANY_REQUESTS = 429,
+  INTERNAL_SERVER_ERROR = 500,
+  NOT_IMPLEMENTED = 501,
+  BAD_GATEWAY = 502,
+  SERVICE_UNAVAILABLE = 503,
 }
 
 export const resetColor = "\x1b[0m";
@@ -66,8 +66,14 @@ export const greenColor = "\x1b[32m";
 export const blueColor = "\x1b[34m";
 export const cyanColor = "\x1b[36m";
 
-export function startupMessage(config: Config, bunServe: Server, routes: Route[]) {
-  const { version } = JSON.parse(fs.readFileSync(import.meta.dir + "/../package.json", "utf-8"))
+export function startupMessage(
+  config: Config,
+  bunServe: Server,
+  routes: Route[]
+) {
+  const { version } = JSON.parse(
+    fs.readFileSync(import.meta.dir + "/../package.json", "utf-8")
+  );
 
   const schema = (config.serveOptions as TLSServeOptions).tls?.key
     ? "https"
@@ -77,8 +83,7 @@ export function startupMessage(config: Config, bunServe: Server, routes: Route[]
   const pidMsg = `PID ...... ${process.pid}`;
   const handlerMsgAndPidMsg = `${handlersMsg}  ${pidMsg}`;
   let successMsg = `ikari v${version}`;
-  if(!version) 
-    successMsg = `ikari`;
+  if (!version) successMsg = `ikari`;
 
   const maxLength = Math.max(
     hostMsg.length,
@@ -100,9 +105,104 @@ export function startupMessage(config: Config, bunServe: Server, routes: Route[]
     .padStart(Math.floor((targetLength - hostMsg.length) / 2) + hostMsg.length)
     .padEnd(targetLength)}${resetColor}│
   │${" ".repeat(maxLength + separatorCount)}│   
-  │${handlerMsgAndPidMsg.padStart(Math.floor((targetLength - handlerMsgAndPidMsg.length) / 2) + handlerMsgAndPidMsg.length).padEnd(targetLength)}│
+  │${handlerMsgAndPidMsg
+    .padStart(
+      Math.floor((targetLength - handlerMsgAndPidMsg.length) / 2) +
+        handlerMsgAndPidMsg.length
+    )
+    .padEnd(targetLength)}│
   └${"─".repeat(maxLength + separatorCount)}┘
 `;
 
   return msg;
+}
+
+//TODO: Tests are missing for this function. We have to test it.
+export function returnContextResponse(ctx: Context) {
+  if (ctx.method === HttpMethod.HEAD || ctx.method === HttpMethod.OPTIONS) {
+    return ctx.getResWithoutBody();
+  }
+  return ctx.res;
+}
+
+export function NotFound(ctx: Context) {
+  if (ctx.method === HttpMethod.HEAD) {
+    return ctx.status(StatusCode.NOT_FOUND).getResWithoutBody();
+  }
+  return ctx.json({ message: "Not Found" }, StatusCode.NOT_FOUND).res;
+}
+
+//TODO: Tests are missing for this function. We have to test it.
+export function getRoutesFromGroups(config: Config, groups: Group[]): Route[] {
+  return groups.reduce(
+    (result: Route[], { prefix, controllers, middlewares }: Group) => {
+      if (prefix) {
+        prefix = createPath(prefix).replace(/\/+$/, "");
+      }
+
+      controllers.forEach((controller: Controller) => {
+        let routes: Route[] = [];
+        if (Reflect.hasMetadata("routes", controller)) {
+          routes = Reflect.getMetadata("routes", controller);
+        } else {
+          routes = Reflect.getMetadata("routes", controller.prototype);
+        }
+
+        routes.forEach((route) => {
+          route.target = controller;
+          if (prefix) route.path = prefix + route.path;
+          if (config.prefix) route.path = config.prefix + route.path;
+          if (middlewares) {
+            route.before = middlewares.concat(route.before);
+          }
+
+          result.push(route);
+        });
+      });
+
+      return result;
+    },
+    []
+  );
+}
+
+//TODO: Tests are missing for this function. We have to test it.
+export function getRoutesFromControllers(
+  config: Config,
+  controllers: Controller[]
+): Route[] {
+  return controllers.reduce((result: Route[], controller: Controller) => {
+    let routes: Route[] = [];
+    if (Reflect.hasMetadata("routes", controller)) {
+      routes = Reflect.getMetadata("routes", controller);
+    } else {
+      routes = Reflect.getMetadata("routes", controller.prototype);
+    }
+
+    routes.forEach((route) => {
+      route.target = controller;
+      if (config.prefix) {
+        route.path = config.prefix + route.path;
+      }
+
+      result.push(route);
+    });
+
+    return result;
+  }, []);
+}
+
+//TODO: Tests are missing for this function. We have to test it.
+export function defaultErrorHandler(err: Errorlike) {
+  return new Response(
+    JSON.stringify({
+      message: err?.message,
+      stack: err?.stack,
+      cause: err?.cause,
+    }),
+    {
+      status: 500,
+      headers: { "Content-Type": "application/json" },
+    }
+  );
 }
